refactor: remove nested exports from features and entities
