--- src/entities/ProductCard/index.ts
@@ -1,2 +1,0 @@
-export { ProductCard } from "./ui";
-export type { TProductCardImage, TProductCardProps } from "./ui";
