--- src/entities/ProductCard/ui/index.ts
@@ -1,2 +1,0 @@
-export { ProductCard } from "./ProductCard";
-export type { TProductCardImage, TProductCardProps } from "./types";
