+++ src/entities/index.ts
@@ -1,2 +1,5 @@
-export { ProductCard } from "./ProductCard";
-export type { TProductCardImage, TProductCardProps } from "./ProductCard";
+export { ProductCard } from "./ProductCard/ui/ProductCard";
+export type {
+  TProductCardImage,
+  TProductCardProps,
+} from "./ProductCard/ui/types";
