--- src/features/ButtonLike/index.ts
@@ -1,1 +1,0 @@
-export { ButtonLike } from "./ui";
