+++ src/features/ButtonLike/ui/ButtonLike.stories.tsx
@@ -6,7 +6,7 @@ import { ButtonLike } from "./ButtonLike";
 
 const productCardArgs = {
   image: {
-    src: "https://avatars.dzeninfra.ru/get-zen_doc/271828/pub_675ecaaece410104475048da_675ef0290659d0610d0cd3f8/scale_1200",
+    src: "https://i.scdn.co/image/7c04200539a69e2a0948836809b484501098c56e",
     alt: "Snowy night cover art",
   },
   title: "ОДИН МАНУЛ",
