--- src/features/ButtonLike/ui/index.ts
@@ -1,2 +1,0 @@
-export { ButtonLike } from "./ButtonLike";
-export type { TButtonLikeProps } from "./types";
